requirements.txt
# Core Web Framework
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
pydantic==2.8.0
pydantic-settings==2.3.0
python-multipart==0.0.9

# AI & LLM
google-generativeai==0.7.0
chromadb==0.5.0
sentence-transformers==3.0.0
langchain==0.2.1
langchain-google-genai==1.0.6

# Telephony
twilio==9.2.0

# Database
sqlalchemy==2.0.31
aiosqlite==0.20.0

# Document Processing
PyMuPDF==1.24.7

# Utilities
httpx==0.27.0
phonenumbers==8.13.40
pydub==0.25.1
pyngrok==7.1.6
